Move MSRC to package Remove MRC specific code from Host.Client
