+++ src/Host/Client/Test/Install/RInstallationTest.cs
@@ -2,14 +2,10 @@
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
 using System;
-using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.IO;
 using FluentAssertions;
-using Microsoft.Common.Core;
 using Microsoft.Common.Core.IO;
-using Microsoft.Common.Core.Shell;
-using Microsoft.Common.Core.Telemetry;
 using Microsoft.Common.Core.Test.Registry;
 using Microsoft.R.Host.Client.Install;
 using Microsoft.UnitTests.Core.FluentAssertions;
@@ -153,52 +149,6 @@ namespace Microsoft.R.Host.Client.Test.Install {
                      new RegistryKeyMock[] {
                              new RegistryKeyMock("8.0.3")
                      }),
-             };
-        }
-
-        [Test]
-        [Category.R.Install]
-        public void MsRClient() {
-            var rClientInstallPath = @"C:\Program Files\Microsoft\R Client\";
-            var rClientRPath = @"C:\Program Files\Microsoft\R Client\R_SERVER\";
-            var tr = new RegistryMock(SimulateRegistryMsRClient(rClientInstallPath, rClientRPath));
-            RInstallation.Registry = tr;
-
-            RInstallation.GetRClientPath().Should().Be(rClientRPath);
-
-            var shell = Substitute.For<ICoreShell>();
-            shell.ShowMessage(Arg.Any<string>(), Arg.Any<MessageButtons>()).Returns(MessageButtons.Yes);
-
-            RInstallation.GetRInstallPath(null, null, shell).Should().Be(rClientRPath);
-            shell.Received(1).ShowMessage(Arg.Any<string>(), Arg.Any<MessageButtons>());
-
-            RInstallation.GetRInstallPath(null, null, shell).Should().Be(rClientRPath);
-            shell.Received(1).ShowMessage(Arg.Any<string>(), Arg.Any<MessageButtons>());
-        }
-
-        private RegistryKeyMock[] SimulateRegistryMsRClient(string rClientInstallPath, string rClientRPath) {
-            return new RegistryKeyMock[] {
-                new RegistryKeyMock(
-                        name: @"SOFTWARE\Microsoft\R Client",
-                        subkeys: null,
-                        valueNames: new string[] {"Path"}, 
-                        values: new string[] {rClientInstallPath}),
-                new RegistryKeyMock(
-                        name: @"SOFTWARE\Microsoft\R Tools\" + Toolset.Version),
-                new RegistryKeyMock(
-                        name: @"SOFTWARE\Microsoft\Microsoft SQL Server\130\sql_shared_mr",
-                        subkeys: null,
-                        valueNames: new string[] {"Path"},
-                        values: new string[] {rClientRPath}),
-                new RegistryKeyMock(
-                       @"SOFTWARE\R-core\R64",
-                        new RegistryKeyMock[] {
-                            new RegistryKeyMock(
-                                name: "3.2.5",
-                                subkeys: null,
-                                valueNames: new string[] {"InstallPath"},
-                                values: new string[] {rClientRPath}),
-                        })
              };
         }
