+++ src/Package/Impl/Commands/R/InstallRClientCommand.cs
@@ -3,6 +3,7 @@
 
 using Microsoft.Common.Core.Shell;
 using Microsoft.R.Host.Client.Install;
+using Microsoft.VisualStudio.R.Package.RClient;
 using Microsoft.VisualStudio.R.Packages.R;
 
 namespace Microsoft.VisualStudio.R.Package.Commands {
@@ -19,7 +20,7 @@ namespace Microsoft.VisualStudio.R.Package.Commands {
         }
 
         protected override void Handle() {
-            var path = RInstallation.GetRClientPath();
+            var path = MicrosoftRClient.GetRClientPath();
             if (!string.IsNullOrEmpty(path)) {
                 _shell.ShowMessage(Resources.Message_RClientIsAlreadyInstalled, MessageButtons.OK);
             } else {
