+++ src/Package/Impl/Options/R/Tools/RToolsSettingsImplementation.cs
@@ -107,7 +107,7 @@ namespace Microsoft.VisualStudio.R.Package.Options.R {
         public RToolsSettingsImplementation() {
             // Default settings. Will be overwritten with actual
             // settings (if any) when settings are loaded from storage
-            RBasePath = RInstallation.GetRInstallPath(null, null, VsAppShell.Current);
+            RBasePath = RInstallation.GetRInstallPath(null, null);
             _workingDirectory = Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments);
         }
