+++ src/Package/Impl/Resources.resx
@@ -779,4 +779,7 @@ Additional information:
   <data name="PleaseRestartVisualStudioAfterRClientSetup" xml:space="preserve">
     <value>Microsoft R Client setup is about to start. Please close Visual Studio now and reopen it when setup is complete.</value>
   </data>
+  <data name="Prompt_MsRClientJustInstalled" xml:space="preserve">
+    <value>Visual Studio detected that Microsoft R Client was recently installed. Would you like to start using R from the Microsoft R Client?</value>
+  </data>
 </root>
