+++ src/Package/Impl/Telemetry/Data/RPackageData.cs
@@ -9,7 +9,6 @@ using System.Security.Cryptography;
 using System.Text;
 using Microsoft.R.Host.Client.Install;
 using Microsoft.R.Support.Settings;
-using Microsoft.VisualStudio.R.Package.Shell;
 
 namespace Microsoft.VisualStudio.R.Package.Telemetry.Data {
     /// <summary>
@@ -22,7 +21,7 @@ namespace Microsoft.VisualStudio.R.Package.Telemetry.Data {
         /// <returns></returns>
         public static IEnumerable<string> GetInstalledPackageHashes(RPackageType packageType) {
 
-            string rInstallPath = RInstallation.GetRInstallPath(RToolsSettings.Current.RBasePath, null, VsAppShell.Current);
+            string rInstallPath = RInstallation.GetRInstallPath(RToolsSettings.Current.RBasePath, null);
             if (!string.IsNullOrEmpty(rInstallPath)) {
                 IEnumerable<string> packageNames = Enumerable.Empty<string>();
                 if (packageType == RPackageType.Base) {
